Added handling for all patterns for extra information
The handling could still use some work but at least there's now a framework in place.
The parsers attribute for the Prereqs class is used to decide what function to use to parse and format based on the type of data the pattern finds. This isn't my favorite way of handling this because it means you need to have two identical strings at two different places (patterns and uOscrape) but I think it's the best way to implement the open/closed principle. I'm also open to other ways if someone has a better suggestion.
We might want to add default values to all the pattern matches, this has pros and cons:
pros:
	- more consistant data format for those who need to read it later
	- by extention possibly much simpler code later
cons:
	- longer more redundant json files
	- slight increase in complexity because you need to set the default values for all attributes in __init__