Simplified get_courses and made it now pass extra info even without prereqs
The get_courses function used to only pass the prerequisite string if the word Prerequisite or Préalable were in the string.
This was causing problems when parsing other information from the extrablock. Then when modifying the function saw that it could be simplified by making the class passed to beautiful soup more specific, and then by using a helper funciton for even further simplicification.